page links and start of navbar
created all of the pages linked in the navbar and added a link back to the home page placed inside the logo.
started working on the navbar but will likely get lazy and move to bootstrap to reduce styling workload. Not too much done but it is starting to resemble a webpage! :D